+++ lib/terraspace/cli/clean/logs.rb
@@ -1,21 +1,19 @@
-class Terraspace::CLI::Logs
-  class Tasks
-    include Terraspace::Util::Sure
-
-    def initialize(options={})
-      @options = options
-      Terraspace.check_project!
+class Terraspace::CLI::Clean
+  class Logs < Base
+    def run
+      action = @options[:truncate] ? "truncate" : "remove"
+      are_you_sure?(action)
+      @options[:truncate] ? truncate : remove
+      logger.info "Logs #{action}d" # IE: Logs truncated or Logs removed
     end
 
     def truncate
-      are_you_sure?("truncate")
       log_files.each do |path|
         File.open(path, "w").close # truncates files
       end
     end
 
     def remove
-      are_you_sure?("remove")
       puts "Removing all files in #{pretty_log_root}/" unless @options[:mute]
       FileUtils.rm_rf(log_root)
       FileUtils.mkdir_p(log_root)
@@ -42,3 +40,4 @@ class Terraspace::CLI::Logs
     end
   end
 end
+